container/kubernetes/onboarding/requirements.txt
GitPython==3.1.3
oisp==0.2.1
PyGithub==1.51
PyYAML==5.3.1
